+++ src/main/java/org/example/controller/NoteController.java
@@ -1,10 +1,10 @@
 package org.example.controller;
 
 import jakarta.validation.Valid;
-import org.example.dto.ApiPrincipal;
-import org.example.dto.Note;
-import org.example.dto.NoteNew;
-import org.example.dto.UserNotes;
+import org.example.dto.user.ApiPrincipal;
+import org.example.dto.note.Note;
+import org.example.dto.note.NoteNew;
+import org.example.dto.user.UserNotes;
 import org.example.service.NoteService;
 import org.example.service.UserService;
 import org.slf4j.Logger;
@@ -12,8 +12,6 @@ import org.slf4j.LoggerFactory;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("/api")
