+++ src/main/java/org/example/controller/UserController.java
@@ -1,7 +1,10 @@
 package org.example.controller;
 
 import jakarta.validation.Valid;
-import org.example.dto.*;
+import org.example.dto.user.User;
+import org.example.dto.user.UserApi;
+import org.example.dto.user.UserCheck;
+import org.example.dto.user.UserNew;
 import org.example.service.UserService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -12,7 +15,7 @@ import org.springframework.web.bind.annotation.*;
 public class UserController {
 
   public final UserService userService;
-  private static final Logger log = LoggerFactory.getLogger(NoteController.class);
+  private static final Logger log = LoggerFactory.getLogger(UserController.class);
 
   public UserController(UserService userService) {
     this.userService = userService;
