+++ src/main/java/org/example/dto/note/Note.java
@@ -1,6 +1,5 @@
-package org.example.dto;
+package org.example.dto.note;
 
-import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
 import java.time.LocalDateTime;
