+++ src/main/java/org/example/dto/note/NoteNew.java
@@ -1,4 +1,4 @@
-package org.example.dto;
+package org.example.dto.note;
 
 import java.time.LocalDateTime;
