+++ src/main/java/org/example/dto/note/NoteResponse.java
@@ -1,4 +1,4 @@
-package org.example.dto;
+package org.example.dto.note;
 
 public record NoteResponse(Long id, String value) {
 }
