+++ src/main/java/org/example/dto/user/ApiPrincipal.java
@@ -1,3 +1,3 @@
-package org.example.dto;
+package org.example.dto.user;
 
 public record ApiPrincipal(Long userId, String email){}
