+++ src/main/java/org/example/dto/user/User.java
@@ -1,4 +1,4 @@
-package org.example.dto;
+package org.example.dto.user;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
