+++ src/main/java/org/example/dto/user/UserApi.java
@@ -1,4 +1,4 @@
-package org.example.dto;
+package org.example.dto.user;
 
 import java.time.LocalDateTime;
