+++ src/main/java/org/example/dto/user/UserNew.java
@@ -1,4 +1,4 @@
-package org.example.dto;
+package org.example.dto.user;
 
 public record UserNew(Long id, String name, String email, String api) {
 }
